searchy/requirements.txt
flask
flask_session
pyyaml

